Update uniauth() function behavior
This commit updates the behavior of the core uniauth() function to
support an optional URL argument. If a call to uniauth() is made with
url=null, then the function will just return null if the session doesn't
exist. This allows users to do away with calling
uniauth_check()/uniauth() in sequence when they really just want to
conditionally obtain the session login info. (This was an anti-pattern.)

Fix issue where get_default_sessid() return value of NULL was not being
handled in calling context which could have resulted in undefined
behavior.